+++ CHANGELOG.md
@@ -62,6 +62,7 @@ All notable changes to this project will be documented in this file.
 - [Multiple] Refactor ZIP importing (#4153 by: HebaruSan)
 - [GUI] Work around OpenFileDialog always showing all shortcuts despite filters (#4168 by: HebaruSan)
 - [Core] Suppress exceptions for non-indexed mods (#4170 by: HebaruSan)
+- [GUI] Label search fixes (#4187 by: HebaruSan)
 
 ### Internal